+++ app/scripts/lib/stores/persistence-manager.ts
@@ -1,9 +1,49 @@
+/**
+ * navigator.locks is a polyfill for the Locks API, which can be removed from
+ * this file after we drop support for older Firefox versions.
+ * navigator.locks is supported in Firefox 102+.
+ */
+import 'navigator.locks';
 import log from 'loglevel';
 import { captureException } from '@sentry/browser';
 import { isEmpty } from 'lodash';
-import { type MetaMaskStateType, MetaMaskStorageStructure } from './base-store';
-import ExtensionStore from './extension-store';
-import ReadOnlyNetworkStore from './read-only-network-store';
+import { KeyringControllerState } from '@metamask/keyring-controller';
+import { MISSING_VAULT_ERROR } from '../../../../shared/constants/errors';
+import { OnboardingControllerState } from '../../controllers/onboarding';
+import { PreferencesControllerState } from '../../controllers/preferences-controller';
+import { AppStateControllerState } from '../../controllers/app-state-controller';
+import { IndexedDBStore } from './indexeddb-store';
+import type {
+  MetaMaskStateType,
+  MetaMaskStorageStructure,
+  BaseStore,
+  MetaData,
+} from './base-store';
+
+export type Backup = {
+  KeyringController?: KeyringControllerState;
+  OnboardingController?: OnboardingControllerState;
+  PreferencesController?: PreferencesControllerState;
+  AppStateController?: AppStateControllerState;
+};
+
+function makeBackup(state?: MetaMaskStateType): Backup {
+  return {
+    KeyringController: state?.KeyringController,
+    OnboardingController: state?.OnboardingController,
+    PreferencesController: state?.PreferencesController,
+    AppStateController: state?.AppStateController,
+  } as Backup;
+}
+
+function hasVault(state?: MetaMaskStateType): state is MetaMaskStateType {
+  const keyringController = state?.KeyringController as
+    | KeyringControllerState
+    | undefined;
+  return Boolean(keyringController?.vault);
+}
+
+const STATE_LOCK = 'state-lock';
 
 /**
  * The PersistenceManager class serves as a high-level manager for handling
@@ -57,58 +97,158 @@ export class PersistenceManager {
    * includes a single key which is 'version' and contains the current version
    * number of the state tree.
    */
-  #metadata?: { version: number };
+  #metadata?: MetaData;
 
   #isExtensionInitialized: boolean = false;
 
-  #localStore: ExtensionStore | ReadOnlyNetworkStore;
+  #localStore: BaseStore;
 
-  constructor({
-    localStore,
-  }: {
-    localStore: ExtensionStore | ReadOnlyNetworkStore;
-  }) {
+  #backupDb: IndexedDBStore;
+
+  #backup?: string;
+
+  #open: boolean = false;
+
+  constructor({ localStore }: { localStore: BaseStore }) {
     this.#localStore = localStore;
+    this.#backupDb = new IndexedDBStore();
   }
 
-  setMetadata(metadata: { version: number }) {
+  async open() {
+    if (!this.#open) {
+      await this.#backupDb.open('metamask-backup', 1);
+      this.#open = true;
+    }
+  }
+
+  setMetadata(metadata: MetaData) {
     this.#metadata = metadata;
   }
 
+  /**
+   * Sets the state in the local store.
+   *
+   * @param state - The state to set in the local store. This should be an object
+   * containing the state data to be stored.
+   * @throws Error if the state is missing or if the metadata is not set before
+   * calling this method.
+   * @throws Error if the local store is not open.
+   * @throws Error if the data persistence fails during the write operation.
+   */
   async set(state: MetaMaskStateType) {
+    await this.open();
+
     if (!state) {
       throw new Error('MetaMask - updated state is missing');
     }
-    if (!this.#metadata) {
+    const meta = this.#metadata;
+    if (!meta) {
       throw new Error('MetaMask - metadata must be set before calling "set"');
     }
-    try {
-      await this.#localStore.set({ data: state, meta: this.#metadata });
-      if (this.#dataPersistenceFailing) {
-        this.#dataPersistenceFailing = false;
-      }
-    } catch (err) {
-      if (!this.#dataPersistenceFailing) {
-        this.#dataPersistenceFailing = true;
-        captureException(err);
-      }
-      log.error('error setting state in local store:', err);
-    } finally {
-      this.#isExtensionInitialized = true;
-    }
+
+    await navigator.locks.request(
+      STATE_LOCK,
+      { mode: 'exclusive' },
+      async () => {
+        try {
+          // atomically set all the keys
+          await this.#localStore.set({
+            data: state,
+            meta,
+          });
+
+          const backup = makeBackup(state);
+          // if we have a vault we can back it up
+          if (backup.KeyringController?.vault) {
+            const stringifiedBackup = JSON.stringify(backup);
+            // and the backup has changed
+            if (this.#backup !== stringifiedBackup) {
+              // save it to the backup DB
+              await this.#backupDb.set(backup);
+              this.#backup = stringifiedBackup;
+            }
+          }
+
+          if (this.#dataPersistenceFailing) {
+            this.#dataPersistenceFailing = false;
+          }
+        } catch (err) {
+          if (!this.#dataPersistenceFailing) {
+            this.#dataPersistenceFailing = true;
+            captureException(err);
+          }
+          log.error('error setting state in local store:', err);
+        } finally {
+          this.#isExtensionInitialized = true;
+        }
+      },
+    );
   }
 
-  async get() {
-    const result = await this.#localStore.get();
+  /**
+   * Retrieves the current state of the local store. If the store is empty,
+   * it returns undefined. If the store is not open, it throws an error.
+   *
+   * @returns The current state of the local store or null if the store is empty.
+   * @throws Error if the vault is missing and a backup vault is found in IndexedDB.
+   * @throws Error if the local store is not open.
+   */
+  async get(): Promise<MetaMaskStorageStructure | null> {
+    await this.open();
 
-    if (isEmpty(result)) {
-      this.#mostRecentRetrievedState = null;
-      return undefined;
-    }
-    if (!this.#isExtensionInitialized) {
-      this.#mostRecentRetrievedState = result;
-    }
-    return result;
+    return await navigator.locks.request(
+      STATE_LOCK,
+      { mode: 'shared' },
+      async () => {
+        const result = await this.#localStore.get();
+
+        // if we don't have a vault
+        if (!hasVault(result?.data)) {
+          // check if we have a backup in IndexedDB. we need to throw an error
+          // so that the user can be told about it. if we don't, carry on as if
+          // everything is fine (it might be, or maybe we lost BOTH the primary
+          // and backup vaults -- yikes!)
+          const backup = await this.getBackup();
+          // this check verifies if we have any keys saved in our backup.
+          // we use this as a sigil to determine if we've ever saved a vault
+          // before.
+          if (Object.values(backup).some((value) => value !== undefined)) {
+            // we've got some data (we haven't checked for a vault, as the
+            // background+UI are responsible for determining what happens now)
+            throw new Error(MISSING_VAULT_ERROR);
+          }
+        }
+
+        if (isEmpty(result)) {
+          this.#mostRecentRetrievedState = null;
+          return undefined;
+        }
+        if (!this.#isExtensionInitialized) {
+          this.#mostRecentRetrievedState = result;
+        }
+        return result;
+      },
+    );
+  }
+
+  async getBackup(): Promise<Backup> {
+    const [
+      KeyringController,
+      OnboardingController,
+      PreferencesController,
+      AppStateController,
+    ] = await this.#backupDb.get([
+      'KeyringController',
+      'OnboardingController',
+      'PreferencesController',
+      'AppStateController',
+    ]);
+    return {
+      KeyringController,
+      OnboardingController,
+      PreferencesController,
+      AppStateController,
+    } as Backup;
   }
 
   get mostRecentRetrievedState() {